+++ src/oai_sgw/s11/s11_common.h
@@ -33,7 +33,7 @@ extern "C" {
 #endif
 
 nw_rc_t s11_ie_indication_generic(uint8_t  ieType,
-                                uint8_t  ieLength,
+                                uint16_t  ieLength,
                                 uint8_t  ieInstance,
                                 uint8_t *ieValue,
                                 void  *arg);
